perf symbol: remove unused 'end' arg in kallsyms parse cb
kallsyms__parse takes a callback that is called on every discovered
symbol. As kallsyms (the kernel interface) does not supply symbol sizes,
the callback was simply called with end=start. Because all the callbacks
are specific to kallsyms__parse, having an end parameter doesn't make
any sense.

Also relocate a comment regarding symbol creation to the callback which
does symbol creation (kallsyms__parse is not in general used to create
symbols).

Signed-off-by: Cody P Schafer <cody@linux.vnet.ibm.com>