perf symbol: only set vmlinux longname & mark loaded if really loaded
dso__load_vmlinux uses the filename passed to it to directly
set the dso long_name, which resulted in a use after free due to
dso__load_vmlinux_path treating 0 symbols as a load failure.

Change dso__load_vmlinux so that finding 0 symbols does not cause it to
consider itself loaded, and do not set long_name in such a case.

Signed-off-by: Cody P Schafer <cody@linux.vnet.ibm.com>