perf symbol: set symtab_type in dso__load_sym
In certain cases, dso__load requires dso->symtab_type to be set prior to
calling it. With the introduction of symsrc*, the symtab_type is know
and stored in a symsrc which is then passed to dso__load_sym.

Change dso__load_sym to use the symtab_type from them symsrc (setting
dso->symtab_type as well).

Setup for later patch
"perf symbol: use both runtime and debug images"

Signed-off-by: Cody P Schafer <cody@linux.vnet.ibm.com>