move debugfs to filesystems menu (fs/Kconfig)
debugfs may have started out as a kernel hacking kind of option.  But,
today, it is an integral part of lots of facilities like tracing that
are far from kernel hacking.  Let's move it out of the very cluttered
"Kernel Hacking" menu and but it next to its real family: other
filesystems like sysfs, configfs, or /proc.

Also, "Debug filesystem" sounds like a debugging option _for_
filesystems code, not a filesystem for debugging.  We also never call
it "the debug filesystem."  We always say "debugfs", so reflect the
fact that we _call_ it debugfs in the menu text.

Signed-off-by: Dave Hansen <dave@linux.vnet.ibm.com>
Acked-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>