tools/perf: don't include libio.h in arch's dwarf-regs.c
Including libio.h causes build failures on uClibc systems (which lack
libio.h).

It appears that libio.h was only included to pull in a
definition for null, so it has been replaced by stddef.h

On powerpc, libio.h was conditionally included, but could be removed
completely as it is unneeded. Also, the included of stdlib.h was changed
to stddef.h (as again, only NULL is needed).

Build tested on powerpc and x86.

Signed-off-by: Cody P Schafer <cody@linux.vnet.ibm.com>