fs/binfmt_misc: avoid accidental exec disable via binfmt_misc loop
On my ubuntu system the update-binfmts command recently went beserk and added a binfmt_misc for executing x86_64 elf files via a static qemu. That static qemu is also an x86_64 elf (as are the majority of the binaries on my x86_64 system). This prevented me from execing any new programs and, due to the lack of an open root shell, made it impossible for me to disable the offending binfmt.

This fix is a hack to disable binfmt_misc handlers when a loop occurs in the hope of saving the system. Ideally, it would also attempt to re-resolve the binfmt that was currently being resolved instead of failing that one and allowing all that follows, but this behavior is much better than the failure of everything that occurs right now.

Signed-off-by: Cody P Schafer <cody@linux.vnet.ibm.com>