mm/page_alloc: don't mistakenly decrease pgdat->nr_zones
In init_currently_empty_zone(), we adjust zone->zone_pgdat->nr_zones
when initing the zone. Previously, the code assumed that the zone
currently being initialized had a higher zone_idx() than all previous
inited zones. This is the case for the current users of
init_currently_empty_zone(), but with DNUMA sometimes we need to
initialize a zone with a lower idx.

For example, take an x86 machine with 2 nodes and 128MB of ram split
into 2 regions (with each node having 1 region). Node 0 will have DMA
and DMA32 zones initialized, while Node 1 will only have DMA32. DNUMA
then comes along and stripes the memory, so some of the DMA memory gets
added to Node 1 and we need to init ZONE_DMA in Node 1.

Signed-off-by: Cody P Schafer <cody@linux.vnet.ibm.com>