mm: zcache: add core files
zcache is a backend for cleancache that takes file pages that are in the process
of being reclaimed and attempts to compress them and store them in a RAM-based
memory pool.
This can result in a significant I/O reduction if system is full
with file pages and, in the case where decompressing from RAM is faster than
reading from the disk, can also improve workload performance.

Signed-off-by: Bob Liu <bob.liu@oracle.com>