perf: rename perf_swevent_event_idx() to perf_event_nop_0() and make extern
Rather an having every pmu that needs a function for .event_idx that
takes an event and unconditionally returns 0 define their own copy,
reuse the one in kernel/events/core.c.

Rename it because we're no longer using it for just software events.
Naming is based on the perf_pmu_nop_*() functions.

Signed-off-by: Cody P Schafer <cody@linux.vnet.ibm.com>